+++ src/main/scala/devices/debug/Debug.scala
@@ -15,7 +15,7 @@ import freechips.rocketchip.util.property._
 import freechips.rocketchip.devices.debug.systembusaccess._
 import freechips.rocketchip.diplomaticobjectmodel.DiplomaticObjectModelAddressing
 import freechips.rocketchip.diplomaticobjectmodel.model._
-import freechips.rocketchip.tile.{LogicalTree, OMRegistry}
+import freechips.rocketchip.tile.LogicalTree
 
 object DsbBusConsts {
   def sbAddrWidth = 12
@@ -1467,8 +1467,8 @@ class TLDebugModule(beatBytes: Int)(implicit p: Parameters) extends LazyModule {
   }
 
   class DebugLogicalTree extends LogicalTree {
-    override def getOMComponents(components: Seq[OMComponent]): Seq[OMComponent] = {
-      device.getOMComponents(OMRegistry.getResourceBindingsMap)
+    override def getOMComponents(resourceBindingsMap: ResourceBindingsMap, components: Seq[OMComponent]): Seq[OMComponent] = {
+      device.getOMComponents(resourceBindingsMap)
     }
   }
